requirements.txt
aiogram==3.17.0
aiosqlite==0.20.0
python-dotenv==1.0.1
anthropic==0.49.0
openai==1.68.2
